PS-807 - Performance improvement

Approved-by: Ed Lyons

+++ src/services/src/Demoulas.ProfitSharing.Services/Internal/ServiceDto/MemberSlice.cs
@@ -8,7 +8,7 @@ internal sealed record MemberSlice
 {
     internal int PsnSuffix { get; init; }
     internal required int BadgeNumber { get; init; }
-    internal long Ssn { get; init; }
+    internal int Ssn { get; init; }
     internal decimal HoursCurrentYear { get; init; }
     internal char EmploymentStatusCode { get; init; }
     internal string? FullName { get; init; } = string.Empty;
